move PostGrid to its own component, reduce nested link styling
