+++ client/src/components/Artist/ArtistPosts.tsx
@@ -2,29 +2,11 @@ import { css } from "@emotion/css";
 import React from "react";
 import { useTranslation } from "react-i18next";
 import { bp } from "../../constants";
-import PostCard from "components/common/PostCard";
 import { useArtistContext } from "state/ArtistContext";
 import SpaceBetweenDiv from "components/common/SpaceBetweenDiv";
 import { FaRss } from "react-icons/fa";
 import { ButtonAnchor } from "components/common/Button";
-import styled from "@emotion/styled";
-
-export const PostGrid = styled.div<{}>`
-  display: grid;
-  grid-template-columns: repeat(3, 31.6%);
-  gap: 4% 2.5%;
-  max-width: 100%;
-  list-style-type: none;
-
-  @media screen and (max-width: ${bp.large}px) {
-    grid-template-columns: repeat(2, 48.75%);
-  }
-
-  @media screen and (max-width: ${bp.medium}px) {
-    grid-template-columns: repeat(1, 100%);
-    gap: 2%;
-  }
-`;
+import PostGrid from "components/Post/PostGrid";
 
 const ArtistPosts: React.FC = () => {
   const { t } = useTranslation("translation", { keyPrefix: "artist" });
@@ -32,6 +14,10 @@ const ArtistPosts: React.FC = () => {
   const {
     state: { artist },
   } = useArtistContext();
+
+  const posts = React.useMemo(() => {
+    return artist?.posts?.map((p) => ({ ...p, artist }));
+  }, [artist]);
 
   if (!artist || artist.posts.length === 0) {
     return null;
@@ -63,13 +49,7 @@ const ArtistPosts: React.FC = () => {
         >
           {artist.posts?.length === 0 && <>{t("noUpdates")}</>}
         </div>
-        <PostGrid as="ul">
-          {artist.posts?.map((p) => (
-            <li key={p.id}>
-              <PostCard p={{ ...p, artist: artist }} />
-            </li>
-          ))}
-        </PostGrid>
+        <PostGrid posts={posts} ariaLabelledBy="artist-navlink-updates" />
       </div>
     </div>
   );
