+++ client/src/components/Home/HomeReleases.tsx
@@ -37,9 +37,6 @@ const Releases = () => {
             padding-bottom: 1rem;
             width: 100%;
 
-            a {
-              color: var(--mi-normal-foreground-color);
-            }
             @media screen and (max-width: ${bp.medium}px) {
               margin-bottom: 0rem;
               padding-top: 0rem;
